encapsulate swizzling
