+++ Sources/MessagingPush/PushHandling/iOSPushEventListener.swift
@@ -11,9 +11,6 @@ protocol PushEventListener: AutoMockable {
     // return nil if the push was not handled by CIO SDK.
     // return true if should diplay the push in foreground.
     func shouldDisplayPushAppInForeground(_ push: PushNotification) -> Bool?
-
-    func newNotificationCenterDelegateSet(_ newDelegate: UNUserNotificationCenterDelegate?)
-    func beginListening()
 }
 
 @available(iOSApplicationExtension, unavailable)
@@ -46,7 +43,6 @@ class iOSPushEventListener: PushEventListener {
     public static let shared = iOSPushEventListener()
 
     // Instances of all dependencies that automated tests can override.
-    private var overrideUserNotificationCenter: UserNotificationCenter?
     private var overrideJsonAdapter: JsonAdapter?
     private var overrideModuleConfig: MessagingPushConfigOptions?
     private var overridePushClickHandler: PushClickHandler?
@@ -56,10 +52,6 @@ class iOSPushEventListener: PushEventListener {
     // Each getter will first check if a test override exists. If so, return that. Otherwise, return an instance from the digraph.
     // It's important that we use do not keep a strong reference to any dependencies or to the digraph. Otherwise, the SDK would crash if the SDK's digraph gets re-initialized and
     // this class tries to access old instances of dependencies.
-    private var userNotificationCenter: UserNotificationCenter? {
-        overrideUserNotificationCenter ?? diGraph?.userNotificationCenter
-    }
-
     private var jsonAdapter: JsonAdapter? {
         overrideJsonAdapter ?? diGraph?.jsonAdapter
     }
@@ -85,12 +77,8 @@ class iOSPushEventListener: PushEventListener {
         SdkInitializedUtilImpl().postInitializedData?.diGraph
     }
 
-    // Make sure that this proxy is held in-memory while the SDK is in memory.
-    private let notificationCenterDelegateProxy = NotificationCenterDelegateProxy()
-
     // Init for testing. Injecting mocks.
-    init(userNotificationCenter: UserNotificationCenter, jsonAdapter: JsonAdapter, moduleConfig: MessagingPushConfigOptions, pushClickHandler: PushClickHandler, pushHistory: PushHistory) {
-        self.overrideUserNotificationCenter = userNotificationCenter
+    init(jsonAdapter: JsonAdapter, moduleConfig: MessagingPushConfigOptions, pushClickHandler: PushClickHandler, pushHistory: PushHistory) {
         self.overrideJsonAdapter = jsonAdapter
         self.overrideModuleConfig = moduleConfig
         self.overridePushClickHandler = pushClickHandler
@@ -99,26 +87,6 @@ class iOSPushEventListener: PushEventListener {
 
     // singleton constructor
     private init() {}
-
-    func beginListening() {
-        guard var userNotificationCenter = userNotificationCenter else {
-            return
-        }
-
-        // Sets up swizzling of `UNUserNotificationCenter.current().delegate` setter to get notified when a new delegate is set on host app.
-        swizzle(
-            forClass: UNUserNotificationCenter.self,
-            original: #selector(setter: UNUserNotificationCenter.delegate),
-            new: #selector(UNUserNotificationCenter.cio_swizzled_setDelegate(delegate:))
-        )
-
-        // This re-assignment triggers NotifiationCenter.delegate setter that we swizzled. We want to run this logic now in case a delegate is already set before the CIO SDK is initialized.
-        userNotificationCenter.currentDelegate = userNotificationCenter.currentDelegate
-    }
-
-    func newNotificationCenterDelegateSet(_ newDelegate: UNUserNotificationCenterDelegate?) {
-        notificationCenterDelegateProxy.newNotificationCenterDelegateSet(newDelegate)
-    }
 
     func onPushAction(_ push: PushNotification, didClickOnPush: Bool) -> Bool {
         guard let pushClickHandler = pushClickHandler,
@@ -185,20 +153,6 @@ class iOSPushEventListener: PushEventListener {
 
         return moduleConfig.showPushAppInForeground
     }
-
-    // Swizzle method convenient when original and swizzled methods both belong to same class.
-    func swizzle(forClass: AnyClass, original: Selector, new: Selector) {
-        guard let originalMethod = class_getInstanceMethod(forClass, original) else { return }
-        guard let swizzledMethod = class_getInstanceMethod(forClass, new) else { return }
-
-        let didAddMethod = class_addMethod(forClass, original, method_getImplementation(swizzledMethod), method_getTypeEncoding(swizzledMethod))
-
-        if didAddMethod {
-            class_replaceMethod(forClass, new, method_getImplementation(originalMethod), method_getTypeEncoding(originalMethod))
-        } else {
-            method_exchangeImplementations(originalMethod, swizzledMethod)
-        }
-    }
 }
 
 // Manually add a getter for the PushEventListener in the digraph.
@@ -208,32 +162,5 @@ extension DIGraph {
     @available(iOSApplicationExtension, unavailable)
     var pushEventListener: PushEventListener {
         iOSPushEventListener.shared
-    }
-}
-
-// This is a bit confusing and makes the code a little more complex. However, it's the most reliable way found to get UNUserNotificationCenter.delegate swizzling to work by using an extension.
-@available(iOSApplicationExtension, unavailable)
-extension UNUserNotificationCenter {
-    // Swizzled method that gets called when `UNUserNotificationCenter.current().delegate` setter called.
-    @objc dynamic func cio_swizzled_setDelegate(delegate: UNUserNotificationCenterDelegate?) {
-        guard let diGraph = SdkInitializedUtilImpl().postInitializedData?.diGraph else {
-            // SDK not initialized yet. We hope that this function never gets called because auto push click handling will not work if this function called.
-
-            // Forward request to the original implementation implementation that we swizzled. So that the app finishes setting UNUserNotificationCenter.delegate.
-            // We have to provide the original delegate object because the SDK has not yet been initialized.
-            cio_swizzled_setDelegate(delegate: delegate)
-
-            return
-        }
-
-        diGraph.logger.debug("New UNUserNotificationCenter.delegate set. Delegate class: \(String(describing: delegate))")
-
-        diGraph.pushEventListener.newNotificationCenterDelegateSet(delegate)
-
-        // Forward request to the original implementation that we swizzled. So that the app finishes setting UNUserNotificationCenter.delegate.
-        //
-        // Instead of providing the given 'delegate', provide CIO SDK's click handler.
-        // This will force our SDK to be the 1 push click handler of the app instead of the given 'delegate'.
-        cio_swizzled_setDelegate(delegate: diGraph.notificationCenterFrameworkAdapter.delegate)
     }
 }
