+++ Sources/MessagingPush/autogenerated/AutoDependencyInjection.generated.swift
@@ -85,7 +85,7 @@ extension DIGraph {
 
     @available(iOSApplicationExtension, unavailable)
     private var newAutomaticPushClickHandling: AutomaticPushClickHandling {
-        AutomaticPushClickHandlingImpl(pushEventListener: pushEventListener, logger: logger)
+        AutomaticPushClickHandlingImpl(notificationCenterFrameworkAdapter: notificationCenterFrameworkAdapter, logger: logger)
     }
 
     // DeepLinkUtil
@@ -107,7 +107,7 @@ extension DIGraph {
     }
 
     private var newNotificationCenterFrameworkAdapter: NotificationCenterFrameworkAdapter {
-        NotificationCenterFrameworkAdapterImpl(pushEventListener: pushEventListener)
+        NotificationCenterFrameworkAdapterImpl(pushEventListener: pushEventListener, userNotificationCenter: userNotificationCenter)
     }
 
     // PushClickHandler
